+++ CHANGELOG.md
@@ -15,7 +15,8 @@ Note: version releases in the 0.x.y range may introduce breaking changes.
 - Fix NullPointerException when event has an empty state (https://github.com/ehrbase/openEHR_SDK/pull/294)
 - Fix issue when template does not contain list of values for DV_ORDINAL (https://github.com/ehrbase/openEHR_SDK/pull/295)
 - Fix issue in AQL regarding LIMIT and OFFSET (https://github.com/ehrbase/openEHR_SDK/pull/296)
-
+- Fix issue while unmarshalling FLAT composition that contains ELEMENT with children DV_CODED_TEXT and DV_TEXT (https://github.com/ehrbase/openEHR_SDK/pull/300)
+- 
 ## 1.16.0
 
 ### Added